requirements.txt
allure_python_commons==2.12.0
concurrent_log_handler==0.9.20
PyAutoGUI==0.9.53
PyMySQL==1.0.2
pytest~=7.2.0
requests==2.28.1
selenium==4.7.2
jsonpath~=0.82
sshtunnel~=0.4.0